% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmissionProb.R
\name{calcEmissionProb}
\alias{calcEmissionProb}
\alias{calcEmissionProb,Poisson,data.frame-method}
\alias{calcEmissionProb,NegBinom,data.frame-method}
\title{Calculate emission probabilities.}
\usage{
calcEmissionProb(distribution, modelData)

\S4method{calcEmissionProb}{Poisson,data.frame}(distribution, modelData)

\S4method{calcEmissionProb}{NegBinom,data.frame}(distribution, modelData)
}
\arguments{
\item{distribution}{A 'Poisson or 'NegBinom' object.}

\item{modelData}{Input data.}
}
\value{
Emission probabilites as a matrix, where each column contains probabilities of one state.
}
\description{
Calculate emission probabilities.
}
\seealso{
\code{\linkS4class{NegBinom}}, \code{\linkS4class{Poisson}}
}
\keyword{internal}
