% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_curve_simulation.R
\name{calculate_TPR_FPR}
\alias{calculate_TPR_FPR}
\title{TPR/FPR}
\usage{
calculate_TPR_FPR(
  positive_vertices,
  negative_vertices,
  true_vertices,
  false_vertices
)
}
\description{
TPR/FPR
}
