% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{color.bar}
\alias{color.bar}
\title{Creates a color bar:}
\usage{
color.bar(
  lut,
  min,
  max = -min,
  nticks = 11,
  ticks = seq(min, max, len = nticks),
  title = ""
)
}
\arguments{
\item{lut}{(function) : The color scale.}

\item{min}{(float) : Minimum value on the color scale}

\item{max}{(float): Maximum value of the color scale}

\item{nticks}{(int) : Number of ticks to have on the color bar.}

\item{ticks}{(vector): The vector of ticks (the user can pass in a vector of custom axes ticks)}

\item{title}{(string): The title of the color bar.}
}
\value{
color_bar (figure) : A color bar.
}
\description{
Taken from http://www.colbyimaging.com/wiki/statistics/color-bars. Creates a color bar from a given color scale and range of tick values.
}
