% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_reconstruction.R
\name{compute_differential_evidence}
\alias{compute_differential_evidence}
\title{Compute a null distribution for function values on a region of a mesh}
\usage{
compute_differential_evidence(
  complex,
  mesh_fcn,
  vertex,
  region_size = 100,
  num_test_regions = 100,
  method = "knn"
)
}
\arguments{
\item{mesh_fcn}{(1 x M vector) the values of the function on the vertices of the mesh.}

\item{vertex}{(int) the index of the desired vertex to analyze on the mesh}

\item{region_size}{(int) the number of vertices to be in the region. We choose the region via knn. The assumption is that the
density if vertices in the mesh is uniform, or at least approximately.}

\item{num_test_regions}{(int) the number of regions to sample from}

\item{method}{(string)  the method to use when creating null regions. We currently support 'area' and 'knn'.}

\item{mesh}{mesh3d object.}
}
\value{
sum of function values for the input region, and a vector of length N of the sum of function values for the random test regions
}
\description{
Input a region on the mesh, and the function finds N other connected regions of the same size, uniformly sampled from the
mesh. The goal is to compare the function values on the input region to those on the N other regions in order to generate the analog of p values.
}
