% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve_simulation.R
\name{compute_roc_curve_cones}
\alias{compute_roc_curve_cones}
\title{Computes the ROC curve by assessing the overlap of selected bands and causal cone.}
\usage{
compute_roc_curve_cones(data, class_1_causal_points, class_2_causal_points,
  distance_to_causal_point = 0.1, rate_values, grid_size, eta = 0.1,
  directions_per_cone, curve_length, truncated = 0, class = 0,
  ball = ball, ball_radius = ball_radius, dir, min_points = 2,
  radius = 2, mode = "grid", subdivision = 3)
}
\arguments{
\item{data}{(list) : Metadata about the simulated shapes (vertex coordinates, etc.)}

\item{class_1_causal_points}{: Vertex indices of causal points for class 1.}

\item{class_2_causal_points}{: Vertex indices of causal points for class 2.}

\item{distance_to_causal_point}{(float) : For interpolated shapes, the distance from a vertex to the causal points to be considered a "causal vertex"}

\item{rate_values}{(vector) : Vector of variable importances for each sub-level set across each direction in a given cone.}

\item{grid_size}{(int) : The fine-ness/granularity of the interpolated shapes.}

\item{eta}{(float) : The kernel shape parameter.}

\item{directions_per_cone}{(int): The number of directions we want generated within each cone.}

\item{curve_length}{(int) : Number of sub-level sets in each EC computation.}

\item{truncated}{(int) : The number of "cuts" to compute TPR/FPR for the ROC curve over. Used to speed up ROC computations.}

\item{class}{(int) : The class of the group of shapes we compute the ROC curve against.}

\item{ball}{(boolean) : Denotes whether or not to compute the EC curves over a ball for uniform measurements}

\item{ball_radius}{(float) : The radius of the bounding ball used if we compute the balled EC curve.}

\item{dir}{(nx3 matrix):  The matrix of directions for which the (S/D) EC curve were computed over.}

\item{min_points}{(int) : Used when type = 'feature'. The mininum number of causal vertices for a sub-level set to be associated with to be considered a causal 'feature'.}

\item{mode}{(string) : The data generation scheme. We currently support 'sphere', 'gaussian_grid", or interpolations (default).}

\item{subdivision}{(int) : The fineness of the sphere meshes (if mode == 'sphere'). We currently use subdivision = 3.}
}
\value{
ROC_dataframe (dataframe) : The ROC curve in Dataframe form.
}
\description{
We compute the ROC curve by assessing the overlap of selected features and causal features. A causal cone here
means to have every direction in a given cone that are associated with a causal vertex to be selected.
We do this for every complex in the data set then average the ROC curves.
}
