% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve_simulation.R
\name{compute_roc_curve_teeth}
\alias{compute_roc_curve_teeth}
\title{Computes the averaged ROC curve of caricatured teeth.}
\usage{
compute_roc_curve_teeth(data_dir1, data_dir2, gamma, class_1_probs,
  class_2_probs, rate_values, directions_per_cone, curve_length,
  directions, truncated = 0, ball_radius = ball_radius, ball = TRUE,
  radius = 0, two_curves = FALSE)
}
\arguments{
\item{data_dir1}{(string) : Location of the first class of meshes.}

\item{data_dir2}{(string) : Location of the second class of meshes.}

\item{gamma}{(float) : The value at which if a vertex has transition probability greater than $\gamma$, then the vertex is considered causal.}

\item{class_1_probs}{(vector/matrix) : The vector/matrix of transition probibilities to propogate the weights for class 1 shapes.}

\item{class_2_probs}{(vector/matrix) : The vector/matrix of transition probibilities to propogate the weights for class 2 shapes.}

\item{rate_values}{(vector) : Vector of variable importances for each sub-level set across each direction in a given cone.}

\item{directions_per_cone}{(int): The number of directions we want generated within each cone.}

\item{curve_length}{(int) : Number of sub-level sets in each EC computation.}

\item{directions}{(nx3 matrix):  The matrix of directions for which the (S/D) EC curve were computed over.}

\item{truncated}{(int) : The number of "cuts" to compute TPR/FPR for the ROC curve over. Used to speed up ROC computations.}

\item{ball_radius}{(float) : The radius of the bounding ball used if we compute the balled EC curve.}

\item{ball}{(boolean) : Denotes whether or not to compute the EC curves over a ball for uniform measurements}

\item{radius}{(int) : The number of sub-level sets "before" and "after" the selected sub-level sets we want to include (during reconstruction).}

\item{two_curves}{(boolean) : Whether or not to compute ROC curves using class specific causal points, or the set of all causal points.}
}
\value{
roc_curve (matrix) : The roc curves for both classes of shapes.
}
\description{
We compute the ROC curve by assessing the overlap of reconstructed vertices and causal vertices.
We do this for every tooth in the directory then average the ROC curves. The user must pass in the locations of the directories for the two meshes,
and also the vector of transition probabilities for the caricaturization procedure.
}
