% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_computation.R
\name{compute_standardized_ec_curve}
\alias{compute_standardized_ec_curve}
\title{Computes the EC curve on a ball}
\usage{
compute_standardized_ec_curve(complex, vertex_function, curve_length,
  first_column_index = FALSE, ball_radius)
}
\arguments{
\item{complex}{(list) : The list containing metadata about the mesh; use \code{process_off_filev3} to obtain this list from an off file, or
\code{convert_complex} to convert existing mesh3d objects to this list.}

\item{vertex_function}{(matrix): A matrix containing the projections of each vertex along a given direction. Computed as the dot product of the vertex onto the
direction on the unit sphere.}

\item{curve_length}{(int): The number of sub-level sets for which to compute the EC curve on in a given direction.}

\item{first_column_index}{(boolean): Specifying the vertex index is included in the vertex function matrix.}

\item{ball_radius}{(float): The radius of the bounding ball.}
}
\value{
ec_curve (nx2 matrix): A matrix containing the EC curve of the given simplicial complex with the first index as the projections for which the EC was computed on.
}
\description{
Computes the EC curve of a mesh on a bounding ball of specified radius $radius$. The filtration steps for which the
(S/D) EC curve are computed is done relative to the bounding ball. For filtrations steps that are computed relative to the shape, use
\code{compute_discrete_ec_curve}. For comparisons with multiple shapes, this is the recommended function.
}
