% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{convert_off_file}
\alias{convert_off_file}
\title{Convert Mesh3d File to List}
\usage{
convert_off_file(mesh)
}
\arguments{
\item{mesh}{(mesh3d): The input mesh to transform to a list.}
}
\value{
complex (list): The mesh converted to list form.
}
\description{
converts a mesh3d file to a list format that's more friendly to our ECT functions.
}
