% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_functions.R
\name{create_comparison_matrix_mult_d}
\alias{create_comparison_matrix_mult_d}
\title{Generate design matrix from meshes in two directories.
 @description Given two classes & directories, we generate the (S/D) EC curves and the associated class labels}
\usage{
create_comparison_matrix_mult_d(
  directory1,
  directory2,
  directions,
  len,
  ball = FALSE,
  ball_radius = 1,
  ec_type = "ECT"
)
}
\arguments{
\item{directory1}{(string): The first directory corresponding to class 1.}

\item{directory2}{(string): The second directory correponding to class 2.}

\item{directions}{(nx3 matrix): The matrix of directions to compute the (S/D) EC curve over.}

\item{len}{(int): The number of sub-level sets to compute the (S/D) EC curve on in each direction.}

\item{ball}{(boolean): Determining whether or not to compute the (S/D) EC curve on a ball for uniform comparisons.}

\item{ball_radius}{(float): The radius of the ball to compute the (S/D) EC on; if you want the curve to be computed relative to the shape, don't touch this parameter.}

\item{ec_type}{(string): What type of EC curve to compute. Currently we support ECT (Euler Characteristic Curve), SECT (Smooth Euler Characteristic Curve)
and DECT (Differentiated Euler Characteristic Curve). We use ECT in the papers.}
}
\value{
final_matrix (matrix): The matrix containing the EC curves of the meshes from the two classes and the associated class labels.
}
\description{
Generate design matrix from meshes in two directories.
 @description Given two classes & directories, we generate the (S/D) EC curves and the associated class labels
}
