% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_directions.R
\name{cross}
\alias{cross}
\title{Cross Product}
\usage{
cross(x, y)
}
\arguments{
\item{x}{(vector): the first (1,3) vector in the cross product.}

\item{y}{(vector): the second (1,3) vector in the cross product.}
}
\value{
a (vector): the cross product of x and y.
}
\description{
Compute the Cross Product
}
