% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_computation.R
\name{differentiate_ec_curve}
\alias{differentiate_ec_curve}
\title{Differentiates an EC curve}
\usage{
differentiate_ec_curve(ec_curve)
}
\arguments{
\item{ec_curve}{(matrix) : The Euler Characteristic Curve from \code{compute_standardized_ec_curve} or \code{compute_discrete_ec_curve}.}
}
\value{
ec_curve (matrix): The differentiated Euler Characteristic Curve (DECT)
}
\description{
Computes the numerical derivative of an Euler Characteristic Curve curve using successive differences.
}
