% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Emission-class.R
\docType{class}
\name{Emission-class}
\alias{Emission-class}
\title{Emission Class}
\description{
The `Emission` class is a generic container for the model used in excess count detection. It defines the probability distributions that describe the emission process in the model.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character string giving the name of the emission.}

\item{\code{mu}}{A numeric matrix where each column stores the mean (`mu`) for the respective hidden state of the model.}

\item{\code{excode_formula}}{An object of class `excodeFormula` specifying the model formula used to describe the emission process.}

\item{\code{glm}}{An object of class `glm` representing the fitted generalized linear model for the emission.}
}}

\seealso{
[stats::glm()] for generalized linear models; [methods::setClass()] for S4 class definitions.
}
