% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Emission-class.R
\name{Emission}
\alias{Emission}
\title{Create an Emission object for excess count detection}
\usage{
Emission(distribution, excode_formula, initial_mu = NULL)
}
\arguments{
\item{distribution}{excodeFamily object.}

\item{excode_formula}{excodeFormula object.}

\item{initial_mu}{Initial estimates of the mean for 'MultiState' models.}
}
\value{
An \code{\linkS4class{Emission}} object.
}
\description{
Create an Emission object for excess count detection
}
