% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Emission-class.R
\docType{class}
\name{EmissionGLMPoisson-class}
\alias{EmissionGLMPoisson-class}
\title{This class defines the Emission function for Poisson GLMs.}
\description{
This class defines the Emission function for Poisson GLMs.
}
\section{Slots}{

\describe{
\item{\code{distribution}}{Name of the distribution ("Poisson").}

\item{\code{excode_formula}}{An \code{\linkS4class{excodeFormula}} object.}
}}

\seealso{
\code{\linkS4class{excodeFormula}}
}
