% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excode-package.R
\docType{package}
\name{excode-package}
\alias{excode}
\alias{excode-package}
\title{excode: Excess Count Detection in Epidemiological Time Series}
\description{
Excess count detection ('excode') in epidemiological time series is an important part of public health surveillance systems. The 'excode' package provides a flexible framework that implements well established approaches to control for seasonality, long-term trends and historic events, but also allows the use of customized models. By combining hidden Markov models and generalized linear models, 'excode' explicitly models normally expected case counts and expected excess case counts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://robert-koch-institut.github.io/excode/}
}

}
\author{
\strong{Maintainer}: Benedikt Zacher \email{ZacherB@rki.de} [copyright holder]

Other contributors:
\itemize{
  \item Ann Christin Vietor [contributor]
}

}
\keyword{internal}
