% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-excodeFamily-class.R
\docType{class}
\name{excodeFamily-class}
\alias{excodeFamily-class}
\title{This class is a generic container for a family of probability distributions}
\description{
This class is a generic container for a family of probability distributions
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the probability distribution.}

\item{\code{params}}{Character vector of name of parameters.}
}}

