% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-excodeFormula-class.R
\docType{class}
\name{excodeFormula-class}
\alias{excodeFormula-class}
\title{This class is a generic container for a formula which specifies the model used for excess count detection.}
\description{
This class is a generic container for a formula which specifies the model used for excess count detection.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the model. Must be one of c("FarringtonNoufaily", "Harmonic", "Custom", "MultiState").}

\item{\code{shared_params}}{Indicates whether model parameters are shared across multiple time series.}

\item{\code{coefficients}}{Coefficients of a fitted model.}

\item{\code{failed_optim}}{TRUE if an error occured during model fitting.}
}}

