% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-excodeModel-class.R
\docType{class}
\name{excodeModel-class}
\alias{excodeModel-class}
\title{This class is a generic container for the model used for outbreak detection}
\description{
This class is a generic container for the model used for outbreak detection
}
\section{Slots}{

\describe{
\item{\code{nStates}}{Number of states in the hidden Markov model.}

\item{\code{emission}}{Emission of hidden Markov model. Includes a excodeFamily and a excodeFormula object.}

\item{\code{transitions}}{Transition proabilities of hidden Markov model.}

\item{\code{transitions_prior}}{Indicates whether a prior for transition probabilities should be used.}

\item{\code{prior_weights}}{Dirichlet prior weights for transition probabilities. This cannot be changed by the user.}

\item{\code{loglik_transitions}}{Prior log likelihood of transition probabilities. Only relevant for model fitting.}

\item{\code{initial_prob}}{Initial state probabilities of the hidden Markov model.}

\item{\code{setBckgState}}{Indicates whether a background state should be computed for model fitting. Background states are set to 0s and time points where the Anscombe residuals of the initial model are < 1.}

\item{\code{id}}{Name of the time series}

\item{\code{method}}{Method for parameter estimation, one of c("supervised", "unsupervised"). Default: "unsupervised".}

\item{\code{converged}}{Logical, indicating whether the EM-algorithm converged.}

\item{\code{niter}}{Maximum number of iterations for model fitting.}

\item{\code{LogLik}}{Log lokelihood of the fitted model.}

\item{\code{AIC}}{Akaike information criterion  of the fitted model.}

\item{\code{BIC}}{Bayesian information criterion  of the fitted model.}

\item{\code{posterior}}{Matrix of posterior probabilities for each time point (rows) and state (columns).}

\item{\code{alpha}}{Alpha (Forward) probabilities of the HMM.}

\item{\code{pval}}{P-value for testing wheter timepoint is in normal or excess state.}

\item{\code{timepoint_fit}}{Numeric time point in timeseries used for fitting.}

\item{\code{timepoint}}{Numeric time point in timeseries used for fitting.}

\item{\code{date}}{Date for each time point.}

\item{\code{observed}}{Vector of observed number of cases.}

\item{\code{population}}{Population size, default: 1.}

\item{\code{error}}{Character string of error message (if an error occured).}
}}

