% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-excodeModel-class.R
\name{excodeModel}
\alias{excodeModel}
\title{Create a model for excess count detection}
\usage{
excodeModel(
  family,
  formula,
  initial_mu = NULL,
  transMat = NULL,
  initProb = NULL,
  transMat_prior = TRUE,
  setBckgState = TRUE
)
}
\arguments{
\item{family}{An \code{\linkS4class{excodeFamily}} object defining the emission distribution
(e.g., Poisson, Negative Binomial).}

\item{formula}{An \code{\linkS4class{excodeFormula}} object specifying the structure of the
model (e.g., time trends, seasonality, ...).}

\item{initial_mu}{Initial estimates of the mean for 'MultiState' models.}

\item{transMat}{Inital transition probabilities.}

\item{initProb}{A numeric vector containing initial state probabilities (probabilities of of states at first time point) of the hidden Markov model.}

\item{transMat_prior}{Logical. Should a prior distribution be used for estimating transition
probabilities? Default is \code{TRUE}.}

\item{setBckgState}{Logical. Should a background state be inferred for model fitting?
Background states are initialized to 0 for time points with Anscombe residuals < 1
from an initial model. Default is \code{TRUE}.}
}
\value{
An object of class \code{\linkS4class{excodeModel}}.
}
\description{
Creates an object of class \code{excodeModel} for modeling excess counts using
a Hidden Markov Model (HMM) with user-defined emission distributions
and formula for modeling observed counts.
}
\examples{
# Initialisation of a mean model without timetrend with Poisson emission

excode_formula_mean <- excodeFormula("Mean", timeTrend = FALSE)
excode_family_pois <- excodeFamily("Poisson")
excodeModel(excode_family_pois, excode_formula_mean)

}
\seealso{
\code{\linkS4class{excodeModel}},
\code{\linkS4class{excodeFamily}},
\code{\linkS4class{excodeFormula}}
}
