% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve_simulation.R
\name{feature_vertex_association}
\alias{feature_vertex_association}
\title{Returns a list that displays the relation of sub-level sets to vertices.}
\usage{
feature_vertex_association(dir, complex, len, ball_radius = 0,
  ball = FALSE)
}
\arguments{
\item{dir}{(nx3 matrix):  The matrix of directions for which the (S/D) EC curve were computed over.}

\item{complex}{(list) : The list containing metadata of the Vertices, Edges, and Faces of the mesh (use process_off_file_v3 to obtain).}

\item{ball_radius}{(float) : The radius of the ball to compute the (S/D) EC on; if you want the curve to be computed relative to the shape, don't touch this parameter.}

\item{ball}{(boolean) : Determining whether or not to compute the (S/D) EC curve on a ball for uniform comparisons.}

\item{rate_vals}{(vector) : Vector of variable importances for each sub-level set across each direction in a given cone.}
}
\value{
vertex_list (list) : The association of vertices to sub-level sets, where the keys are the sub-level set indices, and the values are the vertex indices.
}
\description{
Computes the vertex-sub-level set association.
}
