% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{find_bayesian_variables}
\alias{find_bayesian_variables}
\title{Conducts the Bayesian Feature Selection}
\usage{
find_bayesian_variables(data, param = 0.5, radius = 0, weights = FALSE)
}
\arguments{
\item{data}{(nxm matrix) A nxm matrix containing the covariates and responses.}

\item{param}{(float (0-1)) The cutoff for PIP for including features.}

\item{radius}{(positive integer ). An integer parameter specifying if, and how many neighboring features should be considered for feature selection.
This is done to capture critical points that may be 'close' to the selected feature.}
}
\value{
The output is a vector of indices/features to be selected.
}
\description{
\code{find_bayesian_variables} returns a vector of the selected features/indices of the curve

The function uses the package varbvs to conduct feature selection. See Carbonetto and Stephens 2012.
Once the model is fitted, we set the features who's pip (posterior inclusion probability) is greater than a user specified cutoff (typically 0.1,0.5)
These features are returned along with a window of neighboring features if the user wishes (typically recommended).
}
