% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{find_rate_variables_with_other_sampling_methods}
\alias{find_rate_variables_with_other_sampling_methods}
\title{Derive Rate Values}
\usage{
find_rate_variables_with_other_sampling_methods(gp_data,
  bandwidth = 0.01, type = "Laplace")
}
\arguments{
\item{gp_data}{(matrix) : The design matrix of (S/D) EC curves and the associated class labels}

\item{bandwidth}{(float) : The bandwidth of the Gaussian Kernel used to fit the GPC.}

\item{type}{(string) : The sampling method used. We currently support Laplace's method, Elliptical Slice Sampling, and Expectation Propogation.}
}
\value{
rate_values (nx2 matrix) : The derived variable importance values and the row number denoting which sub-level set it corresponds to.
}
\description{
\code{find_rate_variables_with_other_sampling_methods} returns a vector of variable importances using RATE.
We fit a GPC classifier to the data, draw samples from the latent posterior using one of Lapace Approximation, Elliptical Slice Samping,
or Expectation Propogation. After posterior inference, we fit RATE to derive association measure for each sub-level set of the design matrix.
}
