% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_directions.R
\name{generate_equidistributed_cones}
\alias{generate_equidistributed_cones}
\title{Generate Equidistributed Cones}
\usage{
generate_equidistributed_cones(num_directions, cap_radius, directions_per_cone)
}
\arguments{
\item{num_directions}{(int): The number of equidistributed directions we want on the sphere.}

\item{cap_radius}{(float): The radius of the cones we generate (determines the size of each cone).}

\item{directions_per_cone}{(int): The number of directions we want generated within each cone.}
}
\value{
directions (num_directions*directions_per_cone x 3 matrix): A matrix of equidistributed cones.
}
\description{
Generate cones that are equidistributed about the sphere. We first generate equidistributed points, then use the rodrigues angle formula
to compute cones about those points.
}
