% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_directions.R
\name{generate_equidistributed_points}
\alias{generate_equidistributed_points}
\title{Generate Equidistributed points on a sphere}
\usage{
generate_equidistributed_points(desired_number, N)
}
\arguments{
\item{desired_number}{(int): the desired number of equidistributed points on the 2-sphere.}

\item{N}{(int): the initial number of points that the algorithm will try to generate. If the number of points generated is less than the desired number, the function will increment N}
}
\value{
points (desired_number x 3 matrix): The matrix of equidistributed points in on the sphere.
}
\description{
Generate equidistributed points about a sphere using (insert reference)
}
