% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{generate_gaussian_field}
\alias{generate_gaussian_field}
\title{Place Gaussian Bumps on a Plane}
\usage{
generate_gaussian_field(
  grid_size = 25,
  shared_points,
  causal_points1,
  causal_points2,
  normal_std
)
}
\arguments{
\item{grid_size}{(int): the fineness of the grid.}

\item{shared_points}{(matrix) :rows are shared points between the classes.}

\item{causal_points1}{(matrix) : should be a matrix whose rows are the causal points of class 1}

\item{causal_points2}{(matrix) : should be a matrix whose rows are the causal points of class 2}
}
\value{
complex (list) : Mesh form of the grid.
}
\description{
Place gaussian bumps on a plane at desired locations.
}
