% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{get_heat_colors}
\alias{get_heat_colors}
\title{Reconstructs and finds vertex colors for a set of meshes.}
\usage{
get_heat_colors(dir_name, cuts, pset, comp, colfunc)
}
\arguments{
\item{dir_name}{(string): The input directory containing the meshes to loop over}

\item{cuts}{(int): How many steps to reconstruct when descending the importance values. A higher cut = more fine granularity on the birth times.}

\item{pset}{(list): A list containing the metadata about the analyses (directions, number of sub level sets, directions per cone)}

\item{comp}{(list): A list of elements containing metadata about the variable selection (primarily variable importances)}

\item{colfunc}{(function): The function for the color gradients.}
}
\value{
mesh_list (list) A list containing the colors of the meshes derived from the iterative reconstruction procedure.
}
\description{
Given a set of meshes, color function, and importance values, number of cuts to reconstruct on,the function finds the birth times of all the
vertices and finds vertex colors using the color function provided.
}
