% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{gradGaussKernel}
\alias{gradGaussKernel}
\title{Computes the gradient of the Gauss kernel with respect to the bandwidth}
\usage{
gradGaussKernel(X, bandwidth)
}
\description{
Computes the gradient of the Gauss kernel with respect to the bandwidth
}
