% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-excodeFormula-class.R
\docType{class}
\name{Mean-class}
\alias{Mean-class}
\title{This class is a container for the parameterization of the Mean models.}
\description{
This class is a container for the parameterization of the Mean models.
}
\section{Slots}{

\describe{
\item{\code{timeTrend}}{Indicates whether a time trend should be included in the model.}

\item{\code{timepoints_per_unit}}{Number of time points within the considered time unit (e.g. 52 for weekly observations in a year).}

\item{\code{offset}}{TRUE if an offset should be included in the model.}

\item{\code{formula_bckg}}{A formula which models the 'normal' (background) states.}

\item{\code{formula}}{A formula which models which includes variable(s) to model 'excess' state(s).}
}}

