% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mort_df_germany}
\alias{mort_df_germany}
\title{German all-cause mortality data}
\format{
A tibble with 418 rows and 7 variables:
\describe{
  \item{date}{Week of death}
  \item{observed}{Number of deaths}
  \item{timepoint}{A numeric column counting sequential weeks from 0 to total number of weeks, representing time progression for time-trend modeling}
  \item{sin52}{Computed sinus component used for harmonic modeling}
  \item{cos52}{Computed cosine component used for harmonic modeling}
  \item{bckg_week}{A logical (`TRUE`/`FALSE`) variable marking background mortality weeks, used for estimating model parameters.}
}
}
\usage{
mort_df_germany
}
\description{
A dataset containing the number of weekly deaths in Germany reported to the German Federal Statistical Office \url{https://www.destatis.de/DE/Themen/Gesellschaft-Umwelt/Bevoelkerung/Sterbefaelle-Lebenserwartung/sterbefallzahlen.html}. The variables are as follows:
}
\keyword{datasets}
