% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-excodeFamily-class.R
\docType{class}
\name{NegBinom-class}
\alias{NegBinom-class}
\title{This class is a generic container for a family of Negative Binomial distributions}
\description{
This class is a generic container for a family of Negative Binomial distributions
}
\section{Slots}{

\describe{
\item{\code{nb_size}}{Size parameter of the Negative Binomial distribution. Only relevant if 'MultiState' model is used with a Negative Binomial dsitribution.}
}}

