% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot sts object with excode results}
\usage{
plot_model(excode_model)
}
\arguments{
\item{excode_model}{An object of class \code{\linkS4class{excodeModel}} which specifies the model parameters and structure}
}
\description{
Plot sts object with excode results
}
\examples{

# TODO

}
