% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_functions.R
\name{process_off_file_v3}
\alias{process_off_file_v3}
\title{Read in OFF files to list form}
\usage{
process_off_file_v3(input_dir)
}
\arguments{
\item{input_dir}{(string): The input directory of our off file.}
}
\value{
complex (list): A list containing the Vertex coordinates, edge and face information.
}
\description{
Processing OFF files in a format friendly to our EC curve computations.
}
