% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{rbf_on_grid}
\alias{rbf_on_grid}
\title{Wrapper for RBF Kernel Interpolation}
\usage{
rbf_on_grid(grid_size = 25, func = rbf_gauss, data, eta = 5)
}
\arguments{
\item{grid_size}{(int) The granularity of the grid.}

\item{func}{(function) The interpolation function used. RBF and Inverse quardrtic kernels are supported.}

\item{data}{(nx3 matrix) The data used to interpolate.}

\item{eta}{(float) Varying shape parameter of the kernel}
}
\value{
predictions (matrix) Matrix of Kernel predictions at each point of the grid.
}
\description{
Provides Kernel Interpolated predictions.
}
