% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data_generation.R
\name{rbf_predict}
\alias{rbf_predict}
\title{RBF prediction}
\usage{
rbf_predict(rbf_model, grid)
}
\arguments{
\item{rbf_model}{(list): The RBF kernel model containing metadata bout the model (kernels, rbf weights, interpolation function).}

\item{grid}{(nxn matrix): Grid of points to compute the RBF kernel on.}
}
\value{
predictions (matrix) : The rbf kernel interpolated values.
}
\description{
Computes the RBF kernel interpolation on a provided grid.
}
