% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{reconstruct_faces_on_shape}
\alias{reconstruct_faces_on_shape}
\title{Computes the  birth times of the faces on the shape.}
\usage{
reconstruct_faces_on_shape(
  dir,
  complex,
  rate_vals,
  len,
  cuts = 10,
  cone_size,
  ball_radius,
  ball = TRUE,
  radius = 0
)
}
\arguments{
\item{dir}{(nx3 matrix):  The matrix of directions for which the (S/D) EC curve were computed over.}

\item{complex}{(list) : The list containing metadata of the Vertices, Edges, and Faces of the mesh (use process_off_file_v3 to obtain).}

\item{rate_vals}{(vector) : Vector of variable importances for each sub-level set across each direction in a given cone.}

\item{len}{(int) : The number of sub-level sets which the (S/D) EC curve were computed in each direction.}

\item{cuts}{(int) : How many steps to reconstruct when descending the importance values. A higher cut = more fine granularity on the birth times.}

\item{cone_size}{(int) : The number of directions in each cone.}

\item{ball_radius}{(float) : The radius of the ball to compute the (S/D) EC on; if you want the curve to be computed relative to the shape, don't touch this parameter.}

\item{ball}{(boolean) : Determining whether or not to compute the (S/D) EC curve on a ball for uniform comparisons.}

\item{radius}{(int) : The number of sub-level sets "before" and "after" the selected sub-level sets we want to include.}
}
\value{
face_matrix (matrix) A matrix detailing at which importance value the face was reconstructed, and which "cut" the face was reconstructed relative to the other faces.
}
\description{
Given a specified number of discrete cuts and a vector of importance values, the function computes the
birth times of each face on the shape by iterating through the importance values in descending order and running the reconstruction procedure.
The face birth time is the first importance value at which a vertex corresponding to a face face is reconstructed.
}
