% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarscov2_df}
\alias{sarscov2_df}
\title{SARS-CoV-2 infections in Berlin-Neukölln (Germany)}
\format{
A data.frame with 154 rows and 3 variables:
\describe{
  \item{date}{Day of data collection}
  \item{observed}{Number of cases}
  \item{id}{Name of time series}
}
}
\usage{
sarscov2_df
}
\description{
A dataset containing the daily number of reported SARS-CoV-2 cases from March-July 2020 in Berlin-Neukölln \url{https://robert-koch-institut.github.io/SARS-CoV-2-Infektionen_in_Deutschland/}. The dataset was downloaded on 2024-10-31. The variables are as follows:
}
\keyword{datasets}
