% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shadar_df}
\alias{shadar_df}
\title{Salmonella Hadar cases in Germany 2001-2006}
\format{
A data.frame with 295 rows and 3 variables:
\describe{
  \item{date}{Week of death}
  \item{observed}{Number of cases}
  \item{id}{Name of time series}
}
}
\usage{
shadar_df
}
\description{
A dataset containing the weekly number of reported Salmonella Hadar cases from January 2001 to August 2006 in Germany. The variables are as follows:
}
\keyword{datasets}
