% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_inference.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid Transformation}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{(vector): vector of values to apply the sigmoid transformation}
}
\value{
sigmoid_x (vector): vector of values post sigmoid transformation
}
\description{
\code{sigmoid} Applies the sigmoid transformation
}
