% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snewport_df}
\alias{snewport_df}
\title{Salmonella Newport cases in Germany 2004-2014}
\format{
A data.frame with 528 rows and 3 variables:
\describe{
  \item{date}{Week of death}
  \item{observed}{Number of cases}
  \item{id}{Name of time series}
}
}
\usage{
snewport_df
}
\description{
A dataset containing the weekly number of reported Salmonella Newport cases from January 2004 to February 2014 in Germany. The variables are as follows:
}
\keyword{datasets}
