% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-excodeFormula-class.R
\docType{class}
\name{Splines-class}
\alias{Splines-class}
\title{This class is a container for the parameterization of Spline models. This class is experimental and not recommended for use!}
\description{
This class is a container for the parameterization of Spline models. This class is experimental and not recommended for use!
}
\section{Slots}{

\describe{
\item{\code{df_season}}{Degrees of freedom for spline modeling seasonality. This class uses cubic splines, which are not recommended to model periodic patterns.}

\item{\code{df_trend}}{Degrees of freedom for spline modeling the time trend}

\item{\code{timeTrend}}{Indicates whether a time trend should be included in the model.}

\item{\code{timepoints_per_unit}}{Number of time points within the considered time unit (e.g. 52 for weekly observations in a year).}

\item{\code{offset}}{TRUE if an offset should be included in the model.}

\item{\code{formula_bckg}}{A formula which models the 'normal' (background) states.}

\item{\code{formula}}{A formula which models which includes variable(s) to model 'excess' state(s).}
}}

