% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_computation.R
\name{update_ec_curve}
\alias{update_ec_curve}
\title{Updates the EC curve}
\usage{
update_ec_curve(curve, ec_type)
}
\arguments{
\item{ec_type}{(string): The type of EC curve we want. Currently the SECT, DECT and ECT are supported.}

\item{ec_curve}{(matrix) : The Euler Characteristic Curve from \code{compute_standardized_ec_curve} or \code{compute_discrete_ec_curve}.}
}
\value{
ec_curve (matrix): The (S/D) Euler Characteristic Curve.
}
\description{
Transforms the Euler Characteristic Curve into our desired form.
}
