% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-Emission-class.R
\docType{class}
\name{Emission-class}
\alias{Emission-class}
\title{This class is a generic container for the model used for excess count detection, which defines the probability distributions.}
\description{
This class is a generic container for the model used for excess count detection, which defines the probability distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the emission.}

\item{\code{mu}}{Matrix that stores in each column the mean for the respective state.}

\item{\code{mu0}}{Numeric vector containing the mean for the 'normal' state in 2-state models.}

\item{\code{mu1}}{Numeric vector containing the mean for the 'excess' state in 2-state models. Name of the emission.}
}}

