% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-excodeFormula-class.R
\docType{class}
\name{excodeFormula-class}
\alias{excodeFormula-class}
\title{excodeFormula Class}
\description{
A generic container for a formula specifying the model used for excess count detection.
}
\section{Slots}{

\describe{
\item{\code{name}}{Character; name of the model, one of `c("FarringtonNoufaily","Harmonic","Custom","MultiState")`.}

\item{\code{params}}{Character vector of parameter names included in the model formula.}

\item{\code{intercept}}{Logical; `TRUE` if the model includes an intercept, `FALSE` otherwise.}

\item{\code{coefficients}}{Numeric vector of fitted model coefficients.}

\item{\code{failed_optim}}{Numeric indicator (0/1) showing whether an error occurred during model fitting.}

\item{\code{time_trend}}{Character; type of time trend used in the model (e.g., `"Linear"`, `"Spline1"`, `"Spline2"`, `"None"`).}
}}

