% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-excodeModel-class.R
\docType{class}
\name{excodeModel-class}
\alias{excodeModel-class}
\title{excodeModel Class}
\description{
The `excodeModel` class is a generic container for the model used for outbreak detection based on a hidden Markov model framework.
}
\section{Slots}{

\describe{
\item{\code{nStates}}{Number of states in the hidden Markov model.}

\item{\code{emission}}{Emission component of the hidden Markov model, including an `excodeFamily` and an `excodeFormula` object.}

\item{\code{transitions}}{Transition probabilities of the hidden Markov model.}

\item{\code{transitions_prior}}{Logical indicating whether a prior for transition probabilities should be used.}

\item{\code{prior_weights}}{Dirichlet prior weights for transition probabilities; cannot be changed by the user.}

\item{\code{loglik_transitions}}{Prior log-likelihood of transition probabilities; only relevant for model fitting.}

\item{\code{initial_prob}}{Initial state probabilities of the hidden Markov model.}

\item{\code{setBckgState}}{Logical indicating whether a background state should be computed for model fitting; background states are set to 0 for time points where the Anscombe residuals of the initial model are < 1.}

\item{\code{converged}}{Logical indicating whether the EM algorithm converged.}

\item{\code{niter}}{Number of iterations used for model fitting.}

\item{\code{LogLik}}{Log-likelihood of the fitted model.}

\item{\code{AIC}}{Akaike information criterion of the fitted model.}

\item{\code{BIC}}{Bayesian information criterion of the fitted model.}

\item{\code{posterior}}{Matrix of posterior probabilities for each time point (rows) and state (columns).}

\item{\code{pval}}{P-values for testing whether each time point is in a normal or excess state.}

\item{\code{zscore}}{Standardized Anscombe residuals (z-scores) for each time point.}

\item{\code{timepoint_fit}}{Numeric time point in the time series used for fitting.}

\item{\code{timepoint}}{Numeric time point in the time series.}

\item{\code{date}}{Date corresponding to each time point.}

\item{\code{observed}}{Vector of observed counts or case numbers.}

\item{\code{population}}{Population size; default is 1.}

\item{\code{error}}{Character string containing an error message, if an error occurred during model fitting.}
}}

