% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_excode.R
\name{run_excode}
\alias{run_excode}
\alias{run_excode,sts,excodeModel-method}
\alias{run_excode,list,excodeModel-method}
\alias{run_excode,data.frame,excodeModel-method}
\title{Detect Excess Counts in Epidemiological Time Series}
\usage{
run_excode(
  surv_ts,
  excode_model,
  timepoints,
  learning_type = "unsupervised",
  maxIter = 100,
  verbose = FALSE,
  return_full_model = FALSE,
  past_timepoints_not_incuded = 26,
  time_units_back = 5,
  timepoints_per_unit = 52
)

\S4method{run_excode}{sts,excodeModel}(
  surv_ts,
  excode_model,
  timepoints,
  learning_type = "unsupervised",
  maxIter = 100,
  verbose = FALSE,
  return_full_model = FALSE,
  past_timepoints_not_incuded = 26,
  time_units_back = 5,
  timepoints_per_unit = 52
)

\S4method{run_excode}{list,excodeModel}(
  surv_ts,
  excode_model,
  timepoints,
  learning_type = "unsupervised",
  maxIter = 100,
  verbose = FALSE,
  return_full_model = FALSE,
  past_timepoints_not_incuded = 26,
  time_units_back = 5,
  timepoints_per_unit = 52
)

\S4method{run_excode}{data.frame,excodeModel}(
  surv_ts,
  excode_model,
  timepoints,
  learning_type = "unsupervised",
  maxIter = 100,
  verbose = FALSE,
  return_full_model = FALSE,
  past_timepoints_not_incuded = 26,
  time_units_back = 5,
  timepoints_per_unit = 52
)
}
\arguments{
\item{surv_ts}{A surveillance time series, which can be a \code{data.frame},
an \code{sts} object, or a list of \code{sts} objects.}

\item{excode_model}{An object of class \code{\linkS4class{excodeModel}} which specifies the model parameters and structure}

\item{timepoints}{An \code{integer} or a sequence of integers specifying
the time points for which excess count detection should be performed.}

\item{learning_type}{A \code{character} string indicating the type of learning.
Must be one of \code{c("unsupervised", "semisupervised", "supervised")}.}

\item{maxIter}{An \code{integer} specifying the maximum number of iterations
for the Expectation-Maximization (EM) algorithm. Defaults to \code{100}.}

\item{verbose}{A \code{logical} indicating whether progress should be printed
during execution. Defaults to \code{FALSE}.}

\item{return_full_model}{A \code{logical} indicating whether to return the full
model output. If \code{FALSE} (default), only results for the specified
\code{timepoints} are returned. If \code{TRUE}, the complete time series used
for model fitting is returned.}

\item{past_timepoints_not_incuded}{An \code{integer} specifying the number of past
time points to exclude from initialization. Defaults to \code{26}.}

\item{time_units_back}{Number of years to be used for model fitting.}

\item{timepoints_per_unit}{Number of time points within the considered time unit (e.g. 52 for weekly observations in a year).}
}
\value{
Returns a fitted \code{\linkS4class{excodeModel}} object.
}
\description{
This function fits an `excodeModel` to epidemiological surveillance data
and returns the estimated parameters of an `excodeFamily`. It can handle
single or multiple time series.
}
\examples{

# Example 1: Run with harmonic Poisson model using data.frame
excode_family_pois <- excodeFamily("Poisson")
excode_formula_har <- excodeFormula("Harmonic")
excode_har_pois <- excodeModel(
  excode_family_pois,
  excode_formula_har
)
data(shadar_df)
result_shadar_har <- run_excode(shadar_df, excode_har_pois, 209:295)


# Example 2: Run with sts object (from 'surveillance' package)
\dontrun{
library(surveillance)
data(stsNewport)
result_newport_har <- run_excode(stsNewport, excode_har_pois, 209:295)
}

# TODO need an example different usage of time_units_back and
# timepoints_per_unit which is simpler than the one in vignette

}
\seealso{
\code{\linkS4class{excodeModel}}
}
